+++ src/addon/cc-addon-features.js
@@ -4,6 +4,7 @@ import '../molecules/cc-error.js';
 import { css, html, LitElement } from 'lit-element';
 import { classMap } from 'lit-html/directives/class-map.js';
 import { i18n } from '../lib/i18n.js';
+import { defaultThemeStyles } from '../styles/default-theme.js';
 import { skeletonStyles } from '../styles/skeleton.js';
 
 const cpuSvg = new URL('../assets/cpu.svg', import.meta.url).href;
@@ -145,6 +146,7 @@ export class CcAddonFeatures extends LitElement {
 
   static get styles () {
     return [
+      defaultThemeStyles,
       skeletonStyles,
       // language=CSS
       css`
@@ -155,7 +157,7 @@ export class CcAddonFeatures extends LitElement {
         .feature-list {
           --bdw: 2px;
           --cc-gap: 1rem;
-          --color: #496D93;
+          --color: var(--color-bg-primary);
           --padding: 0.6rem;
         }