+++ src/tcp-redirections/cc-tcp-redirection-form.js
@@ -121,7 +121,7 @@ export class CcTcpRedirectionForm extends LitElement {
         }
 
         .description code {
-          background-color: #f3f3f3;
+          background-color: var(--color-bg-neutral);
           border-radius: 0.25rem;
           font-family: var(--cc-ff-monospace);
           padding: 0.15rem 0.3rem;