+++ src/templates/cc-link.js
@@ -1,6 +1,7 @@
 import { css, html } from 'lit-element';
 import { classMap } from 'lit-html/directives/class-map.js';
 import { ifDefined } from 'lit-html/directives/if-defined.js';
+import { defaultThemeStyles } from '../styles/default-theme';
 
 function isDifferentOrigin (rawUrl) {
   try {
@@ -23,15 +24,16 @@ export const ccLink = (url, content, skeleton = false) => {
 
 // language=CSS
 export const linkStyles = css`
+  ${defaultThemeStyles}
 
   .sanitized-link,
   .cc-link {
-    color: hsl(209, 98%, 40%);
+    color: var(--color-text-primary-highlight);
   }
 
   .sanitized-link:enabled:hover,
   .cc-link:enabled:hover {
-    color: hsl(209, 98%, 25%);
+    color: var(--color-text-primary);
   }
 
   .sanitized-link:focus,