+++ CHANGELOG.md
@@ -6,6 +6,9 @@ title: Changelog
 
 ## Unreleased (????-??-??)
 
+* colors: use variables to specify colors and swap old colors for new contrasted ones.
+* all-form-controls story: display disabled and readonly form controls next to each other to make sure they look the same.
+
 ...
 
 ## 7.10.1 (2022-03-23)